refactor: разделение actions, reducers и api-actions по папкам логически
