+++ src/components/pages/add-review/add-review.jsx
@@ -3,9 +3,9 @@ import MainLayout from "../../layouts/main-layout/main-layout";
 import MovieCard from "../../sections/movie-card/movie-card";
 import {connect} from "react-redux";
 import MoviesList from "../../blocks/movies-list/movies-list";
-import {fetchCurrentMovie} from "../../../store/api-actions";
 import {useParams} from "react-router-dom";
 import Loader from "../../blocks/loader/loader";
+import {fetchCurrentMovie} from "../../../store/movie-data/movie-data-api-actions";
 
 const AddReview = ({currentMovie, isCurrentMovieLoaded, onLoadData}) => {
